fix: Eigen matrix indexing error in bilateral_upsampling.hpp
Convert float indices to int when accessing Eigen matrices.
Newer Eigen versions return IndexedView instead of scalar
when using float indices, causing compile errors.

Fixes:
- Line 112: Add static_cast<int> for matrix access
- Line 119: Add static_cast<int> for vector access

+++ surface/include/pcl/surface/impl/bilateral_upsampling.hpp
@@ -109,14 +109,14 @@ pcl::BilateralUpsampling<PointInT, PointOutT>::performProcessing (PointCloudOut 
             float dx = float (x - x_w),
                 dy = float (y - y_w);
 
-            float val_exp_depth = val_exp_depth_matrix(dx+window_size_, dy+window_size_);
+            float val_exp_depth = val_exp_depth_matrix(static_cast<int>(dx+window_size_), static_cast<int>(dy+window_size_));
 
             float d_color = static_cast<float> (
                 abs (input_->points[y_w * input_->width + x_w].r - input_->points[y * input_->width + x].r) +
                 abs (input_->points[y_w * input_->width + x_w].g - input_->points[y * input_->width + x].g) +
                 abs (input_->points[y_w * input_->width + x_w].b - input_->points[y * input_->width + x].b));
             
-            float val_exp_rgb = val_exp_rgb_vector(d_color);
+            float val_exp_rgb = val_exp_rgb_vector(static_cast<int>(d_color));
 
             if (pcl_isfinite (input_->points[y_w*input_->width + x_w].z))
             {
